ai_pipeline/requirements.txt
tensorflow
numpy
opencv-python
matplotlib
tqdm
scikit-learn
ultralytics
torch
torchvision
Pillow